# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "type": "string"
          },
          "fields": {
            "items": {
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable message.",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FileEntry": {
        "properties": {
          "kind": {
            "enum": [
              "file",
              "dir",
              "symlink",
              "other"
            ],
            "type": "string"
          },
          "mode": {
            "description": "Octal permission bits, e.g. \"644\".",
            "type": "string"
          },
          "mtime": {
            "description": "Modification time, epoch seconds.",
            "type": "integer"
          },
          "name": {
            "description": "Entry name (basename, not a path).",
            "type": "string"
          },
          "size": {
            "description": "Size in bytes.",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "FileListing": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "type": "array"
          },
          "path": {
            "description": "The absolute directory that was listed.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "description": "Correlates the response with server logs.",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "ts"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "Programmatic access with a scoped API key minted at POST /api/v1/account/api-keys. Sent as `Authorization: Bearer \u003ckey\u003e`.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Browser session established by POST /api/v1/auth/login. Mutations additionally require the double-submit CSRF token (cookie hp_csrf echoed in the X-CSRF-Token header) when CSRF is enabled.",
        "in": "cookie",
        "name": "hp_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Control-plane API for the HeroPanel hosting panel (hpd). All responses are wrapped in a `{data, meta}` envelope; errors in a `{error}` envelope. Authenticate with a session cookie (browser) or a bearer API key (programmatic).",
    "license": {
      "name": "See repository"
    },
    "title": "HeroPanel API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sites/{uid}/files": {
      "delete": {
        "description": "Requires the `file.write` permission.",
        "operationId": "api_v1_sites_uid_files_delete_a_file_or_directory",
        "parameters": [
          {
            "description": "Resource UID.",
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query: path=\u003csite-relative path\u003e. Refuses the site root."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or request error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated but missing the required permission"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a file or directory",
        "tags": [
          "Files"
        ],
        "x-required-permission": "file.write"
      },
      "get": {
        "description": "Requires the `file.read` permission.",
        "operationId": "api_v1_sites_uid_files_list_a_directory",
        "parameters": [
          {
            "description": "Resource UID.",
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/FileListing"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Query: path=\u003csite-relative dir\u003e (empty lists the site root). Baremetal sites only."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or request error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated but missing the required permission"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "List a directory",
        "tags": [
          "Files"
        ],
        "x-required-permission": "file.read"
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "apiKey": []
    }
  ],
  "servers": [
    {
      "description": "This hpd instance",
      "url": "/"
    }
  ],
...
}